Set token expires after 72 hours

+++ backend/user_auth/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render
-from rest_framework.authentication import SessionAuthentication, TokenAuthentication
+from rest_framework.authentication import SessionAuthentication
+from .models import ExpiringTokenAuthentication
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.authtoken.models import Token
 from rest_framework.response import Response
@@ -9,20 +10,19 @@ from django.contrib.auth import login, logout, authenticate
 from django.contrib.auth.models import User
 from rest_framework.exceptions import ValidationError
 import json
+from datetime import datetime, timedelta
 
 
 class LogInView(APIView):
     # will follow DEFAULT_AUTHENTICATION_CLASSES in settings.py if unspecified
-    authentication_classes = [TokenAuthentication, SessionAuthentication]
+    authentication_classes = [ExpiringTokenAuthentication, SessionAuthentication]
     permission_classes = []
 
-    def get(self, request, format=None):
-        data = request.data
-        content = {
-            'user': data.get('username', None),
-            'auth': data.get('auth', None),
-        }
-        return Response(content)
+    def get(self, request):
+        user = request.user
+        return Response({"endpoint": "logout-get",
+                         "user": user.username,
+                         "isAuthenticated": user.is_authenticated})
 
     def post(self, request, format=None):
         data = request.data
@@ -34,10 +34,15 @@ class LogInView(APIView):
         user = authenticate(username=username, password=password)
         if user is not None and user.is_active:
             login(request, user)
-            token = Token.objects.get_or_create(user=user)[0].key
+            token, created = Token.objects.get_or_create(user=user)
+
+            if not created:
+                token.created = datetime.utcnow()
+                token.save()
+
             response = {'username': user.username,
                         'result': True,
-                        'authorization': token}
+                        'authorization': token.key}
             return Response(response)
         else:
             raise ValidationError({'result': False,
@@ -45,7 +50,15 @@ class LogInView(APIView):
 
 
 class LogOutView(APIView):
-    def get(self, request, format=None):
-        request.user.auth_token.delete()
-        logout(request)
-        return Response({})
+    def get(self, request):
+        user = request.user
+        return Response({"endpoint": "logout-get",
+                         "user": user.username,
+                         "isAuthenticated": user.is_authenticated})
+
+    def post(self, request, format=None):
+        if request.user.is_authenticated:
+            request.user.auth_token.delete()
+            logout(request)
+            return Response({"result": True})
+        return Response({"result": False, "message": "Unauthenticated user."})
